requirements.txt
Django
Faker
Pillow
django-debug-toolbar
django-polymorphic